Moved controlled decompositions to op_math
