+++ pennylane/ops/op_math/__init__.py
@@ -59,6 +59,14 @@ Controlled Operator Classes
 
     ~ControlledQubitUnitary
 
+Decompositions
+~~~~~~~~~~~~~~
+
+.. autosummary::
+    :toctree: api
+
+    ~ctrl_zyz_decomp
+
 """
 
 from .adjoint import Adjoint, adjoint
@@ -72,6 +80,7 @@ from .prod import Prod, prod
 from .sprod import SProd, s_prod
 from .sum import Sum, op_sum, sum
 from .symbolicop import ScalarSymbolicOp, SymbolicOp
+from .controlled_decompositions import ctrl_decomp_zyz
 
 controlled_qubit_ops = {
     "ControlledQubitUnitary",
